Split tests for `/permissive`
