+++ tests/std/test.lst
@@ -258,6 +258,7 @@ tests\GH_004845_logical_operator_traits_with_non_bool_constant
 tests\GH_004929_internal_tag_constructors
 tests\GH_004930_char_traits_user_specialization
 tests\GH_005090_stl_hardening
+tests\GH_005090_stl_hardening_cxx20_iterators
 tests\GH_005204_regex_collating_ranges
 tests\GH_005236_collate_facet
 tests\GH_005244_regex_escape_sequences
