+++ tests/std/tests/GH_005090_stl_hardening/test.cpp
@@ -5,7 +5,6 @@
 
 #include <array>
 #include <bitset>
-#include <cstddef>
 #include <deque>
 #include <forward_list>
 #include <list>
@@ -21,7 +20,6 @@
 #endif // _HAS_CXX17
 
 #if _HAS_CXX20
-#include <istream>
 #include <iterator>
 #include <ranges>
 #include <span>
@@ -36,200 +34,6 @@
 #include <test_death.hpp>
 
 using namespace std;
-
-#if _HAS_CXX20
-// helper types for common_iterator
-
-struct int_writer {
-    using difference_type = ptrdiff_t;
-
-    int n_{};
-
-    int& operator*() { // non-const
-        return n_;
-    }
-
-    int_writer& operator++() {
-        return *this;
-    }
-    int_writer& operator++(int) {
-        return *this;
-    }
-
-    friend bool operator==(const int_writer&, default_sentinel_t) {
-        return false;
-    }
-};
-
-class lowered_iterator {
-public:
-    using value_type      = int;
-    using difference_type = ptrdiff_t;
-
-    lowered_iterator() = default;
-    explicit lowered_iterator(int* p) : ptr_{p} {}
-
-    int& operator*() const {
-        return *ptr_;
-    }
-
-    int* operator->() const {
-        return ptr_;
-    }
-
-    lowered_iterator& operator++() {
-        ++ptr_;
-        return *this;
-    }
-
-    lowered_iterator operator++(int) {
-        auto old = *this;
-        ++*this;
-        return old;
-    }
-
-    friend bool operator==(const lowered_iterator&, const lowered_iterator&) = default;
-
-    friend bool operator==(const lowered_iterator& i, int* const se) {
-        return i.ptr_ == se;
-    }
-
-private:
-    int* ptr_ = nullptr;
-};
-
-class exceptional_iterator {
-public:
-    using value_type      = int;
-    using difference_type = ptrdiff_t;
-
-    exceptional_iterator() = default;
-    explicit exceptional_iterator(int* p) : ptr_{p} {}
-    exceptional_iterator(const exceptional_iterator& other) : ptr_{other.ptr_} {
-        check_throwing();
-    }
-    exceptional_iterator(exceptional_iterator&) = default;
-
-    exceptional_iterator& operator=(const exceptional_iterator& other) {
-        ptr_ = other.ptr_;
-        check_throwing();
-        return *this;
-    }
-    exceptional_iterator& operator=(exceptional_iterator&&) = default;
-
-    operator lowered_iterator() const {
-        return lowered_iterator{ptr_};
-    }
-
-    int& operator*() const {
-        return *ptr_;
-    }
-
-    int& operator[](const ptrdiff_t n) const {
-        return ptr_[n];
-    }
-
-    int* operator->() const {
-        return ptr_;
-    }
-
-    exceptional_iterator& operator++() {
-        ++ptr_;
-        return *this;
-    }
-
-    exceptional_iterator operator++(int) {
-        auto old = *this;
-        ++*this;
-        return old;
-    }
-
-    exceptional_iterator& operator--() {
-        --ptr_;
-        return *this;
-    }
-
-    exceptional_iterator operator--(int) {
-        auto old = *this;
-        --*this;
-        return old;
-    }
-
-    exceptional_iterator& operator+=(const ptrdiff_t n) {
-        ptr_ += n;
-        return *this;
-    }
-
-    exceptional_iterator& operator-=(const ptrdiff_t n) {
-        ptr_ -= n;
-        return *this;
-    }
-
-    friend bool operator==(const exceptional_iterator& i, const exceptional_iterator& j) {
-        return i.ptr_ == j.ptr_;
-    }
-
-    friend strong_ordering operator<=>(const exceptional_iterator& i, const exceptional_iterator& j) {
-        return i.ptr_ <=> j.ptr_;
-    }
-
-    friend exceptional_iterator operator+(const exceptional_iterator& i, const ptrdiff_t n) {
-        return exceptional_iterator{i.ptr_ + n};
-    }
-    friend exceptional_iterator operator+(const ptrdiff_t n, const exceptional_iterator& i) {
-        return exceptional_iterator{i.ptr_ + n};
-    }
-
-    friend exceptional_iterator operator-(const exceptional_iterator& i, const ptrdiff_t n) {
-        return exceptional_iterator{i.ptr_ - n};
-    }
-    friend ptrdiff_t operator-(const exceptional_iterator& i, const exceptional_iterator& j) {
-        return i.ptr_ - j.ptr_;
-    }
-
-    friend bool operator==(const exceptional_iterator& i, int* const se) {
-        return i.ptr_ == se;
-    }
-
-    friend ptrdiff_t operator-(const exceptional_iterator& i, int* const se) {
-        return i.ptr_ - se;
-    }
-    friend ptrdiff_t operator-(int* const se, const exceptional_iterator& i) {
-        return se - i.ptr_;
-    }
-
-    static void reset_throwing() noexcept {
-        throwing_ = false;
-    }
-
-private:
-    static bool check_throwing() {
-        if (throwing_) {
-            throw bad_variant_access{};
-        }
-        throwing_ = true;
-        return false;
-    }
-
-    int* ptr_ = nullptr;
-
-    static inline bool throwing_ = false;
-};
-
-common_iterator<exceptional_iterator, int*> make_valueless_common_iterator() {
-    using CI = common_iterator<exceptional_iterator, int*>;
-    exceptional_iterator::reset_throwing();
-
-    CI ret{static_cast<int*>(nullptr)};
-    CI src{exceptional_iterator{}};
-    try {
-        ret = move(src);
-    } catch (...) {
-    }
-    exceptional_iterator::reset_throwing();
-    return ret;
-}
-#endif // _HAS_CXX20
 
 // <array>
 void test_array_subscript() {
@@ -571,130 +375,6 @@ void test_string_view_remove_suffix() {
 #endif // _HAS_CXX17
 
 #if _HAS_CXX20
-// <iterator>
-void test_counted_iterator_deref() {
-    int arr[1]{};
-    auto it = counted_iterator<int*>{arr, 1} + 1;
-    (void) *it;
-}
-
-void test_counted_iterator_subscript() {
-    int arr[1]{};
-    counted_iterator<int*> it{arr, 1};
-    (void) it[1];
-}
-
-void test_counted_iterator_iter_move() {
-    int arr[1]{};
-    auto it = counted_iterator<int*>{arr, 1} + 1;
-    (void) ranges::iter_move(it);
-}
-
-void test_counted_iterator_iter_swap() {
-    int arr[1]{};
-    counted_iterator<int*> it{arr, 1};
-    auto jt = it + 1;
-    (void) ranges::iter_swap(it, jt);
-}
-
-void test_counted_iterator_construction() {
-    int arr[1]{};
-    [[maybe_unused]] counted_iterator<int*> it{arr, -1};
-}
-
-void test_counted_iterator_pre_inc() {
-    int arr[1]{};
-    auto it = counted_iterator<int*>{arr, 1} + 1;
-    ++it;
-}
-
-void test_counted_iterator_post_inc_input() {
-    counted_iterator<istream_iterator<int>> it{istream_iterator<int>{}, 0};
-    it++;
-}
-
-void test_counted_iterator_post_inc_forward() {
-    int arr[1]{};
-    auto it = counted_iterator<int*>{arr, 1} + 1;
-    it++;
-}
-
-void test_counted_iterator_advance() {
-    int arr[1]{};
-    counted_iterator<int*> it{arr, 1};
-    it += 2;
-}
-
-void test_counted_iterator_retract() {
-    int arr[1]{};
-    counted_iterator<int*> it{arr, 1};
-    it -= -2;
-}
-
-void test_common_iterator_construct() {
-    auto it = make_valueless_common_iterator();
-    [[maybe_unused]] common_iterator<lowered_iterator, int*> jt{it};
-}
-
-void test_common_iterator_assign() {
-    auto it = make_valueless_common_iterator();
-    common_iterator<lowered_iterator, int*> jt{static_cast<int*>(nullptr)};
-    jt = it;
-}
-
-void test_common_iterator_deref() {
-    common_iterator<int_writer, default_sentinel_t> it{default_sentinel};
-    (void) *it;
-}
-
-void test_common_iterator_deref_const() {
-    common_iterator<counted_iterator<int*>, default_sentinel_t> it{default_sentinel};
-    (void) *it;
-}
-
-void test_common_iterator_arrow() {
-    struct S {
-        int n;
-    };
-    common_iterator<counted_iterator<S*>, default_sentinel_t> it{default_sentinel};
-    (void) it->n;
-}
-
-void test_common_iterator_pre_inc() {
-    common_iterator<counted_iterator<int*>, default_sentinel_t> it{default_sentinel};
-    (void) ++it;
-}
-
-void test_common_iterator_post_inc() {
-    common_iterator<counted_iterator<int*>, default_sentinel_t> it{default_sentinel};
-    (void) ++it;
-}
-
-void test_common_iterator_equality() {
-    auto it = make_valueless_common_iterator();
-    common_iterator<exceptional_iterator, int*> jt{static_cast<int*>(nullptr)};
-    (void) (it == jt);
-}
-
-void test_common_iterator_difference() {
-    auto it = make_valueless_common_iterator();
-    common_iterator<exceptional_iterator, int*> jt{static_cast<int*>(nullptr)};
-    (void) (it - jt);
-}
-
-void test_common_iterator_iter_move() {
-    common_iterator<counted_iterator<int*>, default_sentinel_t> it{default_sentinel};
-    (void) ranges::iter_move(it);
-}
-
-void test_common_iterator_iter_swap() {
-    using CI = counted_iterator<int*>;
-
-    int arr[1]{};
-    common_iterator<CI, default_sentinel_t> it{default_sentinel}, jt{CI{arr, 1}};
-    (void) ranges::iter_swap(it, jt);
-}
-
 // <ranges>
 // ranges::subrange derives from ranges::view_interface, which is hardened.
 void test_ranges_view_interface_subscript() {
@@ -1007,27 +687,6 @@ int main(int argc, char* argv[]) {
 #endif // _HAS_CXX17
 
 #if _HAS_CXX20
-        test_counted_iterator_deref,
-        test_counted_iterator_subscript,
-        test_counted_iterator_iter_move,
-        test_counted_iterator_iter_swap,
-        test_counted_iterator_construction,
-        test_counted_iterator_pre_inc,
-        test_counted_iterator_post_inc_input,
-        test_counted_iterator_post_inc_forward,
-        test_counted_iterator_advance,
-        test_counted_iterator_retract,
-        test_common_iterator_construct,
-        test_common_iterator_assign,
-        test_common_iterator_deref,
-        test_common_iterator_deref_const,
-        test_common_iterator_arrow,
-        test_common_iterator_pre_inc,
-        test_common_iterator_post_inc,
-        test_common_iterator_iter_move,
-        test_common_iterator_iter_swap,
-        test_common_iterator_equality,
-        test_common_iterator_difference,
         test_ranges_view_interface_subscript,
         test_ranges_view_interface_subscript_const,
         test_ranges_view_interface_front,
