+++ tests/std/tests/GH_005090_stl_hardening_cxx20_iterators/env.lst
@@ -1,0 +1,6 @@
+# Copyright (c) Microsoft Corporation.
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+RUNALL_INCLUDE ..\strict_20_matrix.lst
+RUNALL_CROSSLIST
+*	PM_CL="/D_MSVC_STL_HARDENING=1"
